requirements.txt
SQLAlchemy==1.4.23
Flask_SocketIO==5.1.1
python2_secrets==1.0.5
Werkzeug==2.0.1
WTForms==3.0.0a1
Flask_WTF==0.15.1
Flask_SQLAlchemy==2.5.1
Flask==2.0.1
Flask_Login==0.5.0
eventlet==0.33.0
pytz==2021.1
Pillow==9.0.1
python_dateutil==2.8.2
secrets==1.0.2
unicode_slugify==0.1.5
